Added My Jumping
There is a sample scene in Otwo's Folder where I put my basic jumping script. Hold shift and you will jump further. If we use this, I need you to add a camera and make it so player jumps in the cameras forward vector. It's a fairly simple script but HMU if you need help